adjust column purchase_products, add table payments, purchase_product {create & view} page, fix bugs date timezone

+++ database/migrations/2024_08_11_194011_add_column_into_pivot_table_ordered_purchase_products_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Nette\Schema\Schema as SchemaSchema;
 
 return new class extends Migration
 {
@@ -11,17 +12,17 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('ordered_purchase_products', function(Blueprint $table) {
-            $table->dropColumn('sub_total');
-        });
-        Schema::table('ordered_purchase_products', function(Blueprint $table) {
-            $table->integer('sub_total')->after('selling_price')->comment('Total harga qty * harga jual (satuan)');
-        });
-
         Schema::table('purchase_products', function(Blueprint $table) {
-            $table->dropColumn('sub_total');
+            if (Schema::hasColumn('purchase_products', 'sub_total')) {
+                $table->dropColumn('sub_total');
+            }
         });
 
+        Schema::table('ordered_purchase_products', function(Blueprint $table) {
+            if (!Schema::hasColumn('ordered_purchase_products', 'sub_total')) {
+                $table->integer('sub_total')->after('selling_price')->comment('Total harga qty * harga jual (satuan)');
+            }
+        });
     }
 
     /**
@@ -29,12 +30,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('ordered_purchase_products', function(Blueprint $table) {
-            $table->dropColumn('sub_total')->comment('Total harga qty * harga jual (satuan)');
-        });
-
-        Schema::table('purchase_products', function(Blueprint $table) {
-            $table->integer('sub_total')->comment('Total harga qty * harga jual (satuan)');
-        });
+        if (Schema::hasColumn('ordered_purchase_products', 'sub_total')) {
+            Schema::table('ordered_purchase_products', function(Blueprint $table) {
+                $table->dropColumn('sub_total');$table->dropColumn('sub_total')->comment('Total harga qty * harga jual (satuan)');
+            });
+        }
     }
 };
